dashboard tab as a separate module
